InvisiCaps 1.5: Move aux field flags to bottom bits and aux ptr to top bits
A current inefficiency related to the aux field is that in order to
extract the pointer or check flags, both common operations in generated
code, we need to perform pointer shifts and masks that require 64-bit
constants to be materialized into registers using the movabs instruction
due to the use of high bits above bit position 48, which increases code
size and register pressure. There are at least two possible ways to
avoid this:

1) Move the flags somewhere else (e.g. to the aux object). This is part of
   InvisiCaps 2.0 (#19).
2) Alter the aux field representation to allow for simpler instruction
   sequences for extracting parts of the field. Let's call this
   InvisiCaps 1.5.

This patch implements option 2. We move the pointer to the top bits in
order to allow it to be extracted with a shift-right by 16, and move the
flags to the bottom bits to allow them to be tested using instructions
that take immediate operands.

With this patch I observed an overall .text size decrease of 5-10%, and
Fil reports that it improves PizBench performance by 0.7% on his machine.